+++ tests/data/aclew.csv
@@ -1,0 +1,29 @@
+speaker_type,segment_onset,segment_offset,vcm_type,words,syllables,phonemes
+CHI,0,100,C,,,
+CHI,200,300,N,,,
+CHI,400,500,C,,,
+CHI,600,700,N,,,
+CHI,800,900,C,,,
+CHI,1000,1100,N,,,
+CHI,1200,1300,C,,,
+CHI,1400,1500,N,,,
+CHI,1600,1700,C,,,
+CHI,1800,1900,N,,,
+CHI,2000,2100,C,,,
+CHI,2200,2300,N,,,
+CHI,2400,2500,C,,,
+CHI,2600,2700,N,,,
+CHI,2800,2900,C,,,
+CHI,3000,3100,N,,,
+CHI,3200,3300,C,,,
+CHI,3400,3500,N,,,
+CHI,3600,3700,C,,,
+CHI,3800,3900,N,,,
+FEM,0,100,,1.0,2.0,4.0
+FEM,500,600,,1.0,2.0,4.0
+FEM,1000,1100,,1.0,2.0,4.0
+FEM,1500,1600,,1.0,2.0,4.0
+FEM,2000,2100,,1.0,2.0,4.0
+FEM,2500,2600,,1.0,2.0,4.0
+FEM,3000,3100,,1.0,2.0,4.0
+FEM,3500,3600,,1.0,2.0,4.0
